requirements.txt
scikit-learn
numpy
pandas
nltk
pyyaml
tabulate